+++ frontend/src/components/dashboard/BrowseByEra.tsx
