+++ frontend/src/components/dashboard/Footer.tsx
