+++ frontend/src/components/dashboard/Gallery.tsx
