+++ frontend/src/components/dashboard/PopularComics.tsx
