+++ frontend/src/components/dashboard/PopularPublishers.tsx
