+++ frontend/src/components/dashboard/PopularSeries.tsx
