+++ frontend/src/components/dashboard/Publisher.tsx
